SOLR-14591: Move JoinQuery To Its Own File (#1609)
